# CUDA: generalize FP16 fattn vec kernel