This PR adds a FlashAttention kernel that only uses regular FP16 arithmetic and not any tensor core operations by generalizing the kernel I wrote for batch size 1 to larger batch sizes. It works reasonably well for batch sizes <= 8. The target hardware is the NVIDIA P100 and AMD RX 5000/6000 GPUs although the code also makes it possible to run FlashAttention on AMD RX 7000 (with likely much worse performance) and on other Pascal GPUs (with effectively unusable performance). On my RX 6800 the performance changes as follows:

| GPU     | Model           |    n_batch |         fa | test       |              t/s |
| ------- | --------------- | ---------: | ---------: | ---------- | ---------------: |
| RX 6800 | Mistral 7b q4_0 |          1 |          0 | pp 4096    |            29.22 |
| RX 6800 | Mistral 7b q4_0 |          1 |          1 | pp 4096    |            38.37 |
| RX 6800 | Mistral 7b q4_0 |          2 |          0 | pp 4096    |            55.52 |
| RX 6800 | Mistral 7b q4_0 |          2 |          1 | pp 4096    |            79.58 |
| RX 6800 | Mistral 7b q4_0 |          4 |          0 | pp 4096    |            90.93 |
| RX 6800 | Mistral 7b q4_0 |          4 |          1 | pp 4096    |           113.53 |
| RX 6800 | Mistral 7b q4_0 |          8 |          0 | pp 4096    |           132.21 |
| RX 6800 | Mistral 7b q4_0 |          8 |          1 | pp 4096    |           135.40 |
| RX 6800 | Mistral 7b q4_0 |         16 |          0 | pp 4096    |           122.97 |
| RX 6800 | Mistral 7b q4_0 |         16 |          1 | pp 4096    |            90.23 |
| RX 6800 | Mistral 7b q4_0 |         32 |          0 | pp 4096    |           224.92 |
| RX 6800 | Mistral 7b q4_0 |         32 |          1 | pp 4096    |           175.97 |
| RX 6800 | Mistral 7b q4_0 |         64 |          0 | pp 4096    |           365.67 |
| RX 6800 | Mistral 7b q4_0 |         64 |          1 | pp 4096    |           247.05 |
| RX 6800 | Mistral 7b q4_0 |        128 |          0 | pp 4096    |           464.88 |
| RX 6800 | Mistral 7b q4_0 |        128 |          1 | pp 4096    |           317.34 |
| RX 6800 | Mistral 7b q4_0 |        256 |          0 | pp 4096    |           552.33 |
| RX 6800 | Mistral 7b q4_0 |        256 |          1 | pp 4096    |           376.56 |
| RX 6800 | Mistral 7b q4_0 |        512 |          0 | pp 4096    |           534.94 |
| RX 6800 | Mistral 7b q4_0 |        512 |          1 | pp 4096    |           390.29 |
| RX 6800 | Mistral 7b q4_0 |       1024 |          0 | pp 4096    |           532.45 |
| RX 6800 | Mistral 7b q4_0 |       1024 |          1 | pp 4096    |           388.57 |
| RX 6800 | Mistral 7b q4_0 |       2048 |          0 | pp 4096    |           531.10 |
| RX 6800 | Mistral 7b q4_0 |       2048 |          1 | pp 4096    |           387.92 |
| RX 6800 | Mistral 7b q4_0 |       4096 |          0 | pp 4096    |           530.76 |
| RX 6800 | Mistral 7b q4_0 |       4096 |          1 | pp 4096    |           387.56 |

This PR also rearranges the order of some definitions in `ggml-cuda/common.cuh`. This is because on master `NO_DEVICE_CODE` is broken on AMD due to the definitions being in the wrong order. I also implemented some warp reduce functions that didn't work on AMD.

@jdecourval when you worked on https://github.com/ggerganov/llama.cpp/pull/6773 , did you check that the code actually produces correct results? I encountered an issue where due to missing implementations in `ggml-cuda/common.cuh` the compiler would optimize out part of the kernel which resulted in significantly faster but useless code.