@JohannesGaessler 2024-05-03
There is no ROCm FlashAttention implementation on master. The only reason the test is passing is that `NO_DEVICE_CODE` is broken so the kernel does nothing instead of throwing an error.