@8XXD8 2024-05-04
I get an error if I try `main` with `-fa`
Gpu is a Radeon Pro VII `gfx906` arch
Rocm 6.1
```
ggml-cuda/fattn.cu:219: ERROR: HIP kernel flash_attn_vec_ext_f16 has no device code compatible with HIP arch 1300.
:0:rocdevice.cpp            :2879: 4907811378 us: [pid:98972 tid:0x7f7aad2006c0] Callback: Queue 0x7f766fd00000 aborting with error : HSA_STATUS_ERROR_EXCEPTION: An HSAIL operation resulted in a hardware exception. code: 0x1016
```