@JohannesGaessler 2024-05-04
The issue was that the logic for determining when to compile the kernel was incorrect. Does it work now?