@8XXD8 2024-05-04
> The issue was that the logic for determining when to compile the kernel was incorrect. Does it work now?

Now it runs, but the output is broken when  FlashAttention is enabled and the `--batch-size` is 8 or smaller.

**--batch-size 16**
```
<|begin_of_text|><|start_header_id|>user<|end_header_id|>Tell me a joke<|eot_id|><|start_header_id|>assistant<|end_header_id|>

Here's one:

Why couldn't the bicycle stand up by itself?

(wait for it...)

Because it was two-tired!

Hope that made you laugh!<|eot_id|> [end of text]
```
**--batch-size 8**
```
<|begin_of_text|><|start_header_id|>user<|end_header_id|>Tell me a joke<|eot_id|><|start_header_id|>assistant<|end_header_id|> 8 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
```