@JohannesGaessler 2024-05-04
Thank you for the bug report, the issue was that I wrote back the data in the wrong order for batch sizes 2-8 and I didn't notice because those batch sizes were not being used when I tested for correctness. It should now be fixed.