@slaren 2024-05-08
Personally I am not interested about the hardware that benefits from this change, and I don't expect to make any changes to this code, so if you want to maintain another copy of the flash attn kernel that's entirely up to you. It would be nice to keep the same performance with tensor cores as in master, but IMO the difference in overall performance is small enough that it could be ignored.

More results with the latest commit:

| GPU         | Model          |   Model Size [GiB] |   Num. of Parameters | Test   |   t/s master |   t/s cuda-fa-no-tc-5 |   Speedup |
|:------------|:---------------|-------------------:|---------------------:|:-------|-------------:|----------------------:|----------:|
| RTX 3090 Ti | llama 13B Q4_0 |               6.86 |          13015864320 | pp512  |      2942.95 |               2928.26 |      1.00 |
| RTX 3090 Ti | llama 13B Q4_0 |               6.86 |          13015864320 | pp1024 |      2926.21 |               2907.86 |      0.99 |
| RTX 3090 Ti | llama 13B Q4_0 |               6.86 |          13015864320 | pp2048 |      2861.99 |               2842.56 |      0.99 |
| RTX 3090 Ti | llama 13B Q4_0 |               6.86 |          13015864320 | tg128  |        85.20 |                 84.30 |      0.99 |
| RTX 3090 Ti | llama 30B Q4_0 |              17.09 |          32528943616 | pp512  |      1253.18 |               1246.15 |      0.99 |
| RTX 3090 Ti | llama 30B Q4_0 |              17.09 |          32528943616 | pp1024 |      1240.20 |               1235.61 |      1.00 |
| RTX 3090 Ti | llama 30B Q4_0 |              17.09 |          32528943616 | pp2048 |      1217.37 |               1212.61 |      1.00 |
| RTX 3090 Ti | llama 30B Q4_0 |              17.09 |          32528943616 | tg128  |        39.34 |                 39.09 |      0.99 |
| RTX 3090 Ti | llama 7B Q4_0  |               3.56 |           6738415616 | pp512  |      4960.42 |               4954.14 |      1.00 |
| RTX 3090 Ti | llama 7B Q4_0  |               3.56 |           6738415616 | pp1024 |      4921.20 |               4894.45 |      0.99 |
| RTX 3090 Ti | llama 7B Q4_0  |               3.56 |           6738415616 | pp2048 |      4787.11 |               4776.69 |      1.00 |
| RTX 3090 Ti | llama 7B Q4_0  |               3.56 |           6738415616 | tg128  |       138.61 |                135.68 |      0.98 |