@slaren 2024-05-09
I think so.

| GPU         | Model          |   Model Size [GiB] |   Num. of Parameters | Test   |   t/s master |   t/s cuda-fa-no-tc-5 |   Speedup |
|:------------|:---------------|-------------------:|---------------------:|:-------|-------------:|----------------------:|----------:|
| RTX 3090 Ti | llama 13B Q4_0 |               6.86 |          13015864320 | tg128  |        95.56 |                 95.89 |      1.00 |
| RTX 3090 Ti | llama 30B Q4_0 |              17.09 |          32528943616 | tg128  |        42.69 |                 42.69 |      1.00 |
| RTX 3090 Ti | llama 7B Q4_0  |               3.56 |           6738415616 | tg128  |       164.25 |                165.74 |      1.01 |